chore: move shouldEnableClientSideMediaProcessing to upload-media
